[chore][mdsv2] Adjust directory struct.
