+++ src/mdsv2/CMakeLists.txt
@@ -19,7 +19,7 @@ add_subdirectory(filesystem)
 add_subdirectory(mds)
 add_subdirectory(service)
 add_subdirectory(storage)
-add_subdirectory(background_task)
+add_subdirectory(background)
 add_subdirectory(client)
 
 add_library(mdsv2_lib
@@ -35,7 +35,7 @@ target_link_libraries(mdsv2_lib
   mdsv2_filesystem
   mdsv2_storage
   mdsv2_service
-  mdsv2_background_task
+  mdsv2_background
 
   dingofs_utils
