+++ src/mdsv2/background/heartbeat.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mdsv2/service/heartbeat.h"
+#include "mdsv2/background/heartbeat.h"
 
 #include <vector>
